+++ src/editor/fileeditor.c
@@ -87,14 +87,14 @@ start_editing_file(struct descriptor_data *d, const char *fname)
 
     inf = fopen(fname, "r");
     if (inf) {
-        CREATE(target, char, sbuf.st_size + 1);
+        CREATE(target, char, (size_t)sbuf.st_size + 1);
         if (!target) {
             errlog("Couldn't allocate memory to edit %s", fname);
             send_to_desc(d, "Internal error #09384\r\n");
             return;
         }
 
-        if (fread(target, sizeof(char), sbuf.st_size,
+        if (fread(target, sizeof(char), (size_t)sbuf.st_size,
                 inf) != (size_t) sbuf.st_size) {
             send_to_desc(d, "Internal error #42372\r\n");
             free(target);
